+++ src/misc/changelog.ts
@@ -23,6 +23,7 @@ export const changelog = [
 		"Features",
 		"• Gold/lumber properly tracked on leaderboard",
 		"• Sheep are buffed even more when winning",
+		"• Shops are not visible to wolves",
 		"Bug fixes",
 		"• Hidden Saving Farms have correct scaling (1.15 -> 1)",
 		"• Fixed Discord link",