+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ## Features
 - Gold/lumber properly tracked on leaderboard
 - Sheep are buffed even more when winning
+- Shops are not visible to wolves
 
 ## Bug fixes
 - Hidden Saving Farms have correct scaling (1.15 -> 1)